Rename missing components
