+++ pages/_app.tsx
@@ -14,8 +14,8 @@ import { store } from '../src/core/store/store';
 import { getAuthFromStorage } from '../src/core/utils/auth-storage';
 import { parseCookie } from '../src/core/utils/cookie-helpers';
 import * as gtag from '../src/core/utils/gtag';
+import { ContainerNotification } from '../src/stories/components/Notification/Notification';
 import { SEOHead } from '../src/stories/components/SEOHead/SEOHead';
-import { ContainerNotification } from '../src/stories/components/notification-rename/notification-rename';
 import AppLayout from '../src/stories/containers/layout/layout';
 import type { CookiesInterface } from '../src/core/utils/types-helpers';
 import type { EmotionCache } from '@emotion/react';
