+++ src/stories/containers/recognized-delegates/recognized-delegates.tsx
@@ -3,7 +3,7 @@ import { CustomLink } from '@ses/components/CustomLink/CustomLink';
 import { CustomPager } from '@ses/components/CustomPager/CustomPager';
 import DelegateSummary from '@ses/components/DelegateSummary/DelegateSummary';
 import { SEOHead } from '@ses/components/SEOHead/SEOHead';
-import { Tabs } from '@ses/components/tabs-rename/Tabs';
+import { Tabs } from '@ses/components/Tabs/Tabs';
 import { CommentActivityContext } from '@ses/core/context/CommentActivityContext';
 import { BudgetStatus } from '@ses/core/models/dto/core-unit.dto';
 import { toAbsoluteURL } from '@ses/core/utils/url.utils';
