+++ src/stories/containers/transparency-report/useTransparencyReport.tsx
@@ -1,4 +1,4 @@
-import CommentsTab from '@ses/components/tabs-rename/CommentsTab/CommentsTab';
+import CommentsTab from '@ses/components/Tabs/CommentsTab/CommentsTab';
 import { CURRENT_ENVIRONMENT } from '@ses/config/endpoints';
 import { getLastUpdateForBudgetStatement } from '@ses/core/business-logic/core-units';
 import { useAuthContext } from '@ses/core/context/AuthContext';
