+++ src/stories/containers/users/users-manager/useManagerAccountLayout.tsx
@@ -8,7 +8,7 @@ import { fetcher } from '../../../../core/utils/fetcher';
 import { ParenthesisNumber } from '../../transparency-report/transparency-report';
 import { QUERY_USERS } from './user-manager.api';
 import type { UserDTO } from '../../../../core/models/dto/auth.dto';
-import type { TabItem } from '../../../components/tabs-rename/Tabs';
+import type { TabItem } from '../../../components/Tabs/Tabs';
 
 export const useManagerAccountLayout = () => {
   const [FEATURE_AUTH] = useState<boolean>(featureFlags[CURRENT_ENVIRONMENT].FEATURE_AUTH);
